fix: move test files to match project convention (no __tests__ dirs)
Verified: precommit-verify ✓ (6ea8bc6ceeaa9164)
